+++ frontend/src/queries/graphql-queries.graphql
@@ -152,6 +152,11 @@ query ProfModalOverviewData($professorId: Int!, $hasEvals: Boolean!) {
     course_professors {
       course {
         ...CourseModalPrefetchCourseData
+        course_id
+        evaluation_statistic @include(if: $hasEvals) {
+          avg_workload
+          avg_rating
+        }
       }
     }
   }
